+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+- Add the option to hide the borders `FrameConfig::hide_border`
 
 ## 0.11.0
 - Improve `ab_glyph` rendering to properly account for glyph outlines that would have previously been out of bounds